+++ lib/src/widgets/el_tooltip/el_tooltip.dart
@@ -1,7 +1,8 @@
 library el_tooltip;
 
 import 'package:flutter/material.dart';
-import 'package:webviewx_plus/webviewx_plus.dart';
+import 'package:quill_html_editor/src/widgets/webviewx/src/webviewx_plus.dart';
+
 import 'src/arrow.dart';
 import 'src/bubble.dart';
 import 'src/element_box.dart';