+++ lib/src/widgets/el_tooltip/src/position_manager.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+
 import 'element_box.dart';
 import 'enum/el_tooltip_position.dart';
 import 'tooltip_elements_display.dart';