+++ CHANGELOG.md
@@ -1,9 +1,100 @@
-## 2.0.5
+## 2.0.7
+- Added [OnFocusChanged] callback function to listen to editor focus changes
+- Updated [webview_flutter] and dependent packages to latest versions in Webviewx
+- [Breaking Change]Added textStyle, textAlign, padding options to the editor default text and 
+   the hint text, removed defaultFontSize, defaultFontColor fields
+- Fixed - [BUG] Native HTML background is showing when keyboard appears #19
+- Added Table feature, we can now add table, insert rows & columns in the editor with interactive picker 
+- Fixed OverlayState bug #16
+- Other minor enhancements and bug fixes
 
-- Added [defaultFontColor] to the QuillHtmlEditor for custom themes, eg.Dark modes
-- Fixed a bug : Toolbar crashes, when user applies multiple colors and selects the text with multiple colors
+
+## 2.0.6-dev.0.3
+
+- Fixed the Readme file 
+- Added Table feature, we can now add table, insert rows & columns in the editor with interactive picker 
+- Fixed OverlayState bug #16 
+- Other minor enhancements and bug fixes
+
+
+## 2.0.5
+## 2.0.4
+- Fixed Readme file 
+- Added [defaultFontColor] to the QuillHtmlEditor for custom themes, eg.Dark modes 
+- Fixed a bug : Toolbar crashes, when user applies multiple colors and selects the text with multiple colors 
 - Added [toolBarColor] for custom theme, to match the background and dropdown color with the theme.
-- Fixed Readme file
+
+
+## 2.0.3
+- Added capability to change toolbar icons color and size 
+- Added option to change the default font size of the editor 
+- Fixed analysis issues with Flutter 3.7
+
+
+## 2.0.2
+## 2.0.1
+## 2.0.0-beta.1
+- Added onTextChanged listeners
+- Added ability to add custom toolbar buttons
+- Added ability to change the padding of toolbar buttons
+- Fixed bug - OnTextChanged is not firing when text is copied to editor
+- Changed getSelectionRange return type from dynamic to a selection model
+- Fixed bug - removed invisible extra bottom padding on focus
+- Other minor bug fixes
+
+
+## 2.0.0
+## 2.0.0-beta.1
+
+- Breaking changes - Editor toolbar is detached from editor
+- Customisable toolbar
+- Added support for image picking, embedding videos, hyperlink with custom UI
+- Added custom Color picker
+- Fixed a few bugs
+- Added onTextChanged listeners
+- Added custom buttons feature
+
+
+## 1.0.0-dev.5
+
+- Fixed setHtml method breaking format bug
+- Fixed UX bugs
+- Breaking changes - Editor toolbar is detached from editor
+- Customisable toolbar
+- Added support for image picking, embedding videos, hyperlink with custom UI
+- Added custom Color picker
+- Fixed a few bugs
+
+
+## 1.0.0-dev.5
+
+- Fixed setHtml method breaking format bug
+- Fixed UX bugs
+
+
+## 1.0.0-dev.4
+## 1.0.0-dev.3
+## 1.0.0-dev.2
+
+- Updated Readme
+- Added Customisable Toolbar
+- ToolBar detached from editor
+- Added Quill Controller
+- Added Support for Image, Video, Hyperlink
+
+
+## 0.1.5
+
+- Updated the documentation
+- Improved the method to enable/disable editor
+
+
+## 0.1.4
+
+- Fixed Readme
+
+
+